Refactor the backtrace stuff into a proper class.

+++ src/processors/cpu_traits.hxx
@@ -8,7 +8,7 @@ template <typename CPU>
 std::string addr2str(typename cpu_traits<CPU>::AddrType pc) {
     std::ostringstream addr;
 
-    uint16_t addr_width = sizeof (typename cpu_traits<CPU>::AddrType) * 2;
+    size_t addr_width = sizeof (typename cpu_traits<CPU>::AddrType) * 2;
 
     addr << std::setfill('0') << std::setw(addr_width) << std::hex
          << std::uppercase << pc;
